Fix order_vps method in VpsService (#30)
Partially fix the order_vps method is VpsService. It's now
possible to order a Vps without addons. If addons are specified
anyway, an NotImplementedError will be raised.

The signature error was caused by the addons not being added to
the OrderedDict in transip.client.Client._build_signature_message.
The _build_signature_message method only adds list items if they
are an instance of SudsObject.